+++ tests/Unit/Exception/AliasNameAndServiceNameCannotBeTheSameExceptionTest.php
@@ -7,9 +7,7 @@ namespace Tests\Unit\Exception;
 use Ghostwriter\Container\Container;
 use Ghostwriter\Container\Exception\AliasNameAndServiceNameCannotBeTheSameException;
 use Ghostwriter\Container\Interface\ContainerInterface;
-use Ghostwriter\Container\Interface\Service\DefinitionInterface;
-use Ghostwriter\Container\Service\Definition\ComposerExtraDefinition;
-use Ghostwriter\Container\Service\Provider\ComposerDefinitionProvider;
+use Ghostwriter\Container\Service\Provider\ComposerServiceProvider;
 use Ghostwriter\Container\Service\Provider\ContainerProvider;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\CoversClassesThatImplementInterface;
@@ -18,11 +16,9 @@ use Tests\Unit\AbstractTestCase;
 use Throwable;
 
 #[CoversClass(AliasNameAndServiceNameCannotBeTheSameException::class)]
-#[CoversClass(ComposerExtraDefinition::class)]
 #[CoversClass(Container::class)]
 #[CoversClass(ContainerProvider::class)]
-#[CoversClass(ComposerDefinitionProvider::class)]
-#[CoversClassesThatImplementInterface(DefinitionInterface::class)]
+#[CoversClass(ComposerServiceProvider::class)]
 #[CoversClassesThatImplementInterface(ContainerInterface::class)]
 final class AliasNameAndServiceNameCannotBeTheSameExceptionTest extends AbstractTestCase
 {