+++ tests/Unit/Exception/InvalidArgumentExceptionTest.php
@@ -8,10 +8,8 @@ use Ghostwriter\Container\Container;
 use Ghostwriter\Container\Exception\InvalidArgumentException;
 use Ghostwriter\Container\Interface\ContainerExceptionInterface;
 use Ghostwriter\Container\Interface\ContainerInterface;
-use Ghostwriter\Container\Interface\Service\DefinitionInterface;
 use Ghostwriter\Container\Interface\Service\ExtensionInterface;
-use Ghostwriter\Container\Service\Definition\ComposerExtraDefinition;
-use Ghostwriter\Container\Service\Provider\ComposerDefinitionProvider;
+use Ghostwriter\Container\Service\Provider\ComposerServiceProvider;
 use Ghostwriter\Container\Service\Provider\ContainerProvider;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\CoversClassesThatImplementInterface;
@@ -23,24 +21,13 @@ use Throwable;
 use function sprintf;
 
 #[CoversClass(InvalidArgumentException::class)]
-#[CoversClass(ComposerExtraDefinition::class)]
 #[CoversClass(Container::class)]
 #[CoversClass(ContainerProvider::class)]
-#[CoversClass(ComposerDefinitionProvider::class)]
+#[CoversClass(ComposerServiceProvider::class)]
 #[CoversClassesThatImplementInterface(ContainerInterface::class)]
 #[CoversClassesThatImplementInterface(ContainerExceptionInterface::class)]
-#[CoversClassesThatImplementInterface(DefinitionInterface::class)]
 final class InvalidArgumentExceptionTest extends AbstractTestCase
 {
-    /** @throws Throwable */
-    public function testContainerDefine(): void
-    {
-        $this->assertException(InvalidArgumentException::class);
-        $this->expectExceptionMessage(self::class);
-
-        $this->container->define(self::class);
-    }
-
     /** @throws Throwable */
     public function testContainerExtend(): void
     {