+++ tests/Unit/Exception/ShouldNotHappenExceptionTest.php
@@ -8,22 +8,18 @@ use Ghostwriter\Container\Container;
 use Ghostwriter\Container\Exception\ShouldNotHappenException;
 use Ghostwriter\Container\Interface\ContainerExceptionInterface;
 use Ghostwriter\Container\Interface\ContainerInterface;
-use Ghostwriter\Container\Interface\Service\DefinitionInterface;
-use Ghostwriter\Container\Service\Definition\ComposerExtraDefinition;
-use Ghostwriter\Container\Service\Provider\ComposerDefinitionProvider;
+use Ghostwriter\Container\Service\Provider\ComposerServiceProvider;
 use Ghostwriter\Container\Service\Provider\ContainerProvider;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\CoversClassesThatImplementInterface;
 use Tests\Unit\AbstractTestCase;
 
 #[CoversClass(ShouldNotHappenException::class)]
-#[CoversClass(ComposerExtraDefinition::class)]
 #[CoversClass(Container::class)]
 #[CoversClass(ContainerProvider::class)]
-#[CoversClass(ComposerDefinitionProvider::class)]
+#[CoversClass(ComposerServiceProvider::class)]
 #[CoversClassesThatImplementInterface(ContainerInterface::class)]
 #[CoversClassesThatImplementInterface(ContainerExceptionInterface::class)]
-#[CoversClassesThatImplementInterface(DefinitionInterface::class)]
 final class ShouldNotHappenExceptionTest extends AbstractTestCase
 {
     public function testExample(): void