+++ tests/Unit/Service/Provider/AbstractProviderTest.php
@@ -7,9 +7,8 @@ namespace Tests\Unit\Service\Provider;
 use Ghostwriter\Container\Container;
 use Ghostwriter\Container\Interface\BuilderInterface;
 use Ghostwriter\Container\Interface\ContainerInterface;
-use Ghostwriter\Container\Service\Definition\ComposerExtraDefinition;
 use Ghostwriter\Container\Service\Provider\AbstractProvider;
-use Ghostwriter\Container\Service\Provider\ComposerDefinitionProvider;
+use Ghostwriter\Container\Service\Provider\ComposerServiceProvider;
 use Ghostwriter\Container\Service\Provider\ContainerProvider;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\UsesClass;
@@ -18,33 +17,32 @@ use Tests\Fixture\Provider\RegisterProviderMock;
 use Tests\Unit\AbstractTestCase;
 
 #[CoversClass(AbstractProvider::class)]
-#[UsesClass(ComposerDefinitionProvider::class)]
-#[UsesClass(ComposerExtraDefinition::class)]
+#[UsesClass(ComposerServiceProvider::class)]
 #[UsesClass(Container::class)]
 #[UsesClass(ContainerProvider::class)]
 final class AbstractProviderTest extends AbstractTestCase
 {
     public function testBootIsNoOp(): void
     {
-        $provider = new BootProviderMock();
+        $bootProviderMock = new BootProviderMock();
         $container = $this->createMock(ContainerInterface::class);
 
         $container->expects(self::never())
             ->method('get')
             ->seal();
 
-        self::assertNull($provider->boot($container));
+        self::assertNull($bootProviderMock->boot($container));
     }
 
     public function testRegisterIsNoOp(): void
     {
-        $provider = new RegisterProviderMock();
+        $registerProviderMock = new RegisterProviderMock();
         $builder = $this->createMock(BuilderInterface::class);
 
         $builder->expects(self::never())
             ->method('alias')
             ->seal();
 
-        self::assertNull($provider->register($builder));
+        self::assertNull($registerProviderMock->register($builder));
     }
 }